restructure c++ projects to have headers and source in different directures.
change #includes to avoid using relative dirs above current directory
