+++ ElectricalCircuitSimulator/SimulationEngine/include/Capacitor.h
