+++ ElectricalCircuitSimulator/SimulationEngine/include/CircuitComponent.h
