+++ ElectricalCircuitSimulator/SimulationEngine/include/GroundedVoltageSource.h
