+++ ElectricalCircuitSimulator/SimulationEngine/include/Inductor.h
