+++ ElectricalCircuitSimulator/SimulationEngine/include/LinearCircuit.h
