+++ ElectricalCircuitSimulator/SimulationEngine/include/Resistor.h
