+++ ElectricalCircuitSimulator/SimulationEngine/src/Capacitor.cpp
