+++ ElectricalCircuitSimulator/SimulationEngine/src/CircuitComponent.cpp
