+++ ElectricalCircuitSimulator/SimulationEngine/src/GroundedVoltageSource.cpp
