+++ ElectricalCircuitSimulator/SimulationEngine/src/Inductor.cpp
