+++ ElectricalCircuitSimulator/SimulationEngine/src/LinearCircuit.cpp
