+++ ElectricalCircuitSimulator/SimulationEngine/src/Resistor.cpp
