+++ ElectricalCircuitSimulator/SimulationEngineWrapper/SimulationEngineWrapper.vcxproj
@@ -70,16 +70,16 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
-    <ClInclude Include="SimulationEngineWrapper.h" />
-  </ItemGroup>
-  <ItemGroup>
-    <ClCompile Include="AssemblyInfo.cpp" />
-    <ClCompile Include="SimulationEngineWrapper.cpp" />
-  </ItemGroup>
-  <ItemGroup>
     <ProjectReference Include="..\SimulationEngine\SimulationEngine.vcxproj">
       <Project>{fc6b0e6c-92b1-49b0-a562-dd53fe59e4a0}</Project>
     </ProjectReference>
+  </ItemGroup>
+  <ItemGroup>
+    <ClInclude Include="include\SimulationEngineWrapper.h" />
+  </ItemGroup>
+  <ItemGroup>
+    <ClCompile Include="src\AssemblyInfo.cpp" />
+    <ClCompile Include="src\SimulationEngineWrapper.cpp" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
