+++ ElectricalCircuitSimulator/SimulationEngineWrapper/SimulationEngineWrapper.vcxproj.filters
@@ -11,15 +11,15 @@
     </Filter>
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="SimulationEngineWrapper.h">
+    <ClInclude Include="include\SimulationEngineWrapper.h">
       <Filter>Header Files</Filter>
     </ClInclude>
   </ItemGroup>
   <ItemGroup>
-    <ClCompile Include="SimulationEngineWrapper.cpp">
+    <ClCompile Include="src\AssemblyInfo.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
-    <ClCompile Include="AssemblyInfo.cpp">
+    <ClCompile Include="src\SimulationEngineWrapper.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
   </ItemGroup>
