+++ ElectricalCircuitSimulator/SimulationEngineWrapper/src/AssemblyInfo.cpp
