+++ ElectricalCircuitSimulator/SimulationEngineWrapper/src/SimulationEngineWrapper.cpp
